helpers: add formatting helper
This commit adds formatting helpers.
We previously had one formatting helper for "Moves".
The formatting library is quite powerful so we might as well use it a
little more.

An example of this is generic enum formatting. We can now simply paste
an enum into the formatter and it will return the static name of the
enum, pretty cool.
